Create Vuex module for Request / View / Edit Study flows
This further simplifies the main store, which now mainly contains top-level UI stuff (e.g. modal, drawer, toast), map, View Data, and authentication.
