+++ web/components/FcActionBottomContinueToConfirm.vue
@@ -18,7 +18,7 @@ export default {
     isSupervisor() {
       return Object.prototype.hasOwnProperty.call(this.$route.query, 'isSupervisor');
     },
-    ...mapState(['studyRequest']),
+    ...mapState('requestStudy', ['studyRequest']),
   },
   validations: ValidationsStudyRequest.validations,
   methods: {
