+++ web/store/index.js
@@ -13,24 +13,14 @@ import {
   Status,
 } from '@/lib/Constants';
 import { debounce } from '@/lib/FunctionUtils';
-import { STUDY_DUPLICATE, STUDY_IRRELEVANT_TYPE } from '@/lib/i18n/ConfirmDialog';
 import DateTime from '@/lib/time/DateTime';
+import requestStudy from '@/web/store/modules/requestStudy';
 import trackRequests from '@/web/store/modules/trackRequests';
 
 Vue.use(Vuex);
 
 const MAX_PER_CATEGORY = 10;
 const TIMEOUT_TOAST = 10000;
-
-function makeStudy(studyType) {
-  return {
-    studyType,
-    daysOfWeek: [2, 3, 4],
-    duration: 24,
-    hours: 'ROUTINE',
-    notes: '',
-  };
-}
 
 function makeItemsCountsActive() {
   const itemsCountsActive = {};
@@ -52,6 +42,7 @@ const clearToastDebounced = debounce((commit) => {
   commit('clearToast');
 }, TIMEOUT_TOAST);
 
+// TODO: DRY with requestStudy module
 function studyRequestEstimatedDeliveryDate(now, studyRequest) {
   if (studyRequest === null) {
     return null;
@@ -70,6 +61,7 @@ function studyRequestEstimatedDeliveryDate(now, studyRequest) {
 
 export default new Vuex.Store({
   modules: {
+    requestStudy,
     trackRequests,
   },
   // TODO: organize state below
@@ -102,11 +94,6 @@ export default new Vuex.Store({
     filterDayOfWeek: [...Array(7).keys()],
     // REQUESTS
     requestReasons: [],
-    // ACTIVE STUDY REQUEST
-    studyRequest: null,
-    studyRequestComments: [],
-    studyRequestLocation: null,
-    studyRequestUser: null,
     // query that will appear in the search bar
     locationQuery: '',
     // DRAWER
@@ -203,23 +190,6 @@ export default new Vuex.Store({
       });
     },
     // ACTIVE STUDY REQUEST
-    studyRequestEstimatedDeliveryDate(state) {
-      const { now, studyRequest } = state;
-      if (studyRequest === null) {
-        return null;
-      }
-      return studyRequestEstimatedDeliveryDate(now, studyRequest);
-    },
-    studyRequestMinDueDate(state) {
-      const { now, studyRequest } = state;
-      if (studyRequest === null) {
-        return null;
-      }
-      if (studyRequest.priority === 'URGENT') {
-        return now;
-      }
-      return now.plus({ months: 2 });
-    },
     studyTypesRelevantToLocation(state) {
       const countTypesAll = COUNT_TYPES.map(({ value }) => value);
       if (state.location === null) {
@@ -241,23 +211,6 @@ export default new Vuex.Store({
       }
       return countTypesAll
         .filter(value => value !== 'TMC' && value !== 'RESCU');
-    },
-    studyTypesWithWarnings(state, getters) {
-      const studyTypesSelected = new Set();
-      if (state.studyRequest !== null) {
-        state.studyRequest.studies.forEach(({ studyType: value }) => {
-          studyTypesSelected.add(value);
-        });
-      }
-      return COUNT_TYPES.map(({ label, value }) => {
-        let warning = null;
-        if (studyTypesSelected.has(value)) {
-          warning = STUDY_DUPLICATE.getModalOptions({ label });
-        } else if (!getters.studyTypesRelevantToLocation.includes(value)) {
-          warning = STUDY_IRRELEVANT_TYPE.getModalOptions({ label });
-        }
-        return { label, value, warning };
-      });
     },
   },
   mutations: {
@@ -323,70 +276,6 @@ export default new Vuex.Store({
     setFilterDayOfWeek(state, filterDayOfWeek) {
       Vue.set(state, 'filterDayOfWeek', filterDayOfWeek);
       Vue.set(state, 'itemsCountsActive', makeItemsCountsActive());
-    },
-    // ACTIVE STUDY REQUEST
-    clearStudyRequest(state) {
-      Vue.set(state, 'studyRequest', null);
-    },
-    setStudyRequest(state, studyRequest) {
-      Vue.set(state, 'studyRequest', studyRequest);
-    },
-    setStudyRequestLocation(state, studyRequestLocation) {
-      Vue.set(state, 'studyRequestLocation', studyRequestLocation);
-    },
-    setNewStudyRequest(state, studyTypes) {
-      const { location, now } = state;
-      const dueDate = now.plus({ months: 3 });
-      const {
-        centrelineId,
-        centrelineType,
-        lng,
-        lat,
-      } = location;
-      const geom = {
-        type: 'Point',
-        coordinates: [lng, lat],
-      };
-      const studies = studyTypes.map(makeStudy);
-      const studyRequest = {
-        serviceRequestId: null,
-        priority: 'STANDARD',
-        assignedTo: null,
-        dueDate,
-        reasons: [],
-        ccEmails: [],
-        centrelineId,
-        centrelineType,
-        geom,
-        studies,
-      };
-      Vue.set(state, 'studyRequest', studyRequest);
-    },
-    addStudyToStudyRequest(state, studyType) {
-      const item = makeStudy(studyType);
-      state.studyRequest.studies.push(item);
-    },
-    removeStudyFromStudyRequest(state, i) {
-      state.studyRequest.studies.splice(i, 1);
-    },
-    setStudyRequestMeta(state, { key, value }) {
-      Vue.set(state.studyRequest, key, value);
-    },
-    setStudyMeta(state, { i, key, value }) {
-      Vue.set(state.studyRequest.studies[i], key, value);
-    },
-    // STUDY REQUEST COMMENTS
-    setStudyRequestComments(state, studyRequestComments) {
-      Vue.set(state, 'studyRequestComments', studyRequestComments);
-    },
-    setStudyRequestCommentUsers(state, studyRequestCommentUsers) {
-      Vue.set(state, 'studyRequestCommentUsers', studyRequestCommentUsers);
-    },
-    addStudyRequestComment(state, comment) {
-      state.studyRequestComments.unshift(comment);
-    },
-    removeStudyRequestComment(state, i) {
-      state.studyRequestComments.splice(i, 1);
     },
     // DRAWER
     setDrawerOpen(state, drawerOpen) {
@@ -518,45 +407,6 @@ export default new Vuex.Store({
       return result;
     },
     // STUDY REQUESTS
-    async fetchStudyRequest({ commit, dispatch }, { id }) {
-      const url = `/requests/study/${id}`;
-      const studyRequest = await apiFetch(url);
-      commit('setStudyRequest', studyRequest);
-
-      const {
-        centrelineId,
-        centrelineType,
-      } = studyRequest;
-
-      const promiseComments = await dispatch('fetchStudyRequestComments', { studyRequest });
-
-      const centrelineIdsAndTypes = [{ centrelineId, centrelineType }];
-      const promiseLocations = dispatch('fetchLocationsFromCentreline', centrelineIdsAndTypes);
-
-      const [
-        studyRequestComments,
-        studyRequestLocations,
-      ] = await Promise.all([promiseComments, promiseLocations]);
-
-      const key = centrelineKey(centrelineType, centrelineId);
-      const studyRequestLocation = studyRequestLocations.get(key);
-      commit('setStudyRequestLocation', studyRequestLocation);
-
-      let subjects = new Set();
-      studyRequestComments.forEach(({ userSubject }) => {
-        subjects.add(userSubject);
-      });
-      subjects = Array.from(subjects);
-      const studyRequestCommentUsers = await dispatch('fetchUsersBySubjects', subjects);
-      commit('setStudyRequestCommentUsers', studyRequestCommentUsers);
-
-      return {
-        studyRequest,
-        studyRequestComments,
-        studyRequestCommentUsers,
-        studyRequestLocation,
-      };
-    },
     async saveStudyRequest({ commit, state }, { isSupervisor, studyRequest }) {
       const { now } = state;
       const estimatedDeliveryDate = studyRequestEstimatedDeliveryDate(now, studyRequest);
@@ -585,57 +435,6 @@ export default new Vuex.Store({
         },
       });
       return studyRequestNew;
-    },
-    // STUDY REQUEST COMMENTS
-    async saveStudyRequestComment({ commit, state }, { studyRequest, comment }) {
-      const { id: studyRequestId } = studyRequest;
-      const url = `/requests/study/${studyRequestId}/comments`;
-      const data = {
-        ...comment,
-      };
-      const options = {
-        method: 'POST',
-        csrf: state.auth.csrf,
-        data,
-      };
-      const commentNew = await apiFetch(url, options);
-      commit('addStudyRequestComment', commentNew);
-    },
-    async fetchStudyRequestComments({ commit }, { studyRequest }) {
-      const { id: studyRequestId } = studyRequest;
-      const studyRequestComments = await apiFetch(`/requests/study/${studyRequestId}/comments`);
-      commit('setStudyRequestComments', studyRequestComments);
-      return studyRequestComments;
-    },
-    async updateStudyRequestComment({ state }, { studyRequest, comment }) {
-      const { id: commentId } = comment;
-      const { id: studyRequestId } = studyRequest;
-      const url = `/requests/study/${studyRequestId}/comments/${commentId}`;
-      const data = {
-        ...comment,
-      };
-      const options = {
-        method: 'PUT',
-        csrf: state.auth.csrf,
-        data,
-      };
-      return apiFetch(url, options);
-    },
-    async deleteStudyRequestComment({ commit, state }, { studyRequest, comment }) {
-      const { id: commentId } = comment;
-      const i = state.studyRequestComments.findIndex(({ id }) => id === commentId);
-      if (i === -1) {
-        return { success: false };
-      }
-      commit('removeStudyRequestComment', i);
-
-      const { id: studyRequestId } = studyRequest;
-      const url = `/requests/study/${studyRequestId}/comments/${commentId}`;
-      const options = {
-        method: 'DELETE',
-        csrf: state.auth.csrf,
-      };
-      return apiFetch(url, options);
     },
     // USERS
     async fetchUsersBySubjects(_, subjects) {
